requirements.txt
annotated-types==0.7.0
awscli==1.36.38
beautifulsoup4==4.12.3
blinker==1.9.0
botocore==1.35.97
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
deptree==0.0.12
distlib==0.3.9
docutils==0.16
easyocr==1.7.2
filelock==3.16.1
Flask==3.1.0
fsspec==2024.2.0
gdown==5.2.0
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.158.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
grpcio==1.69.0
grpcio-status==1.69.0
httplib2==0.22.0
idna==3.10
imageio==2.36.1
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==1.0.1
lazy_loader==0.4
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
ninja==1.11.1.3
numpy==2.2.1
opencv-python-headless==4.10.0.84
packaging==24.2
pillow==11.1.0
platformdirs==4.3.6
proto-plus==1.25.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyclipper==1.3.0.post6
pydantic==2.10.5
pydantic_core==2.27.2
pyparsing==3.2.1
PySocks==1.7.1
python-bidi==0.6.3
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
rsa==4.7.2
s3transfer==0.10.4
scikit-image==0.25.0
scipy==1.15.1
shapely==2.0.6
six==1.17.0
soupsieve==2.6
stripe==11.4.1
sympy==1.13.1
tifffile==2025.1.10
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
virtualenv==20.28.1
Werkzeug==3.1.3
zipp==3.21.0
